+++ demo/src/main/java/flu/demo/table_class/DemoApplication.java
@@ -1,4 +1,4 @@
-package flu.demo;
+package flu.demo.table_class;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
