+++ demo/src/main/java/flu/demo/table_class/Status.java
@@ -1,4 +1,4 @@
-package flu.demo;
+package flu.demo.table_class;
 
 
 import jakarta.persistence.Column;
