+++ demo/src/main/java/flu/demo/table_class/Unit.java
@@ -1,4 +1,4 @@
-package flu.demo;
+package flu.demo.table_class;
 
 
 import jakarta.persistence.Column;
@@ -13,19 +13,19 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-@Table(name="Position")
-public class Position {
+@Table(name="Unit")
+public class Unit {
 
     @Id
-    @Column(name="position_id")
-    private int position_id;
-
-
-    @Column(name="position")
-    private String position;
-
-
     @Column(name="unit_id")
     private int unit_id;
 
+    @Column(name="unit_name")
+    private String unit_name;
+
+    @Column(name="unit_code")
+    private String unit_code;
+
+    @Column(name="department_id")
+    private int department_id;
 }
